削除の確認のアラートを追加

+++ ShoppingList/Screens/Shopping/ShoppingDetail/ShoppingDetailScreen.swift
@@ -55,16 +55,38 @@ struct ShoppingDetailScreen: View {
                 isEditScreen.toggle()
             }
             Button(R.string.button.delete(), role: .destructive) {
-                viewModel.delete()
-                self.dismiss()
+                viewModel.setAlert(type: .confirm, message: R.string.alertMessage.deleteConfirm())
             }
         }
         .sheet(isPresented: $isEditScreen) {
             AddShoppingScreen(viewModel: AddShoppingViewModel(mode: .update,
                                                               shoppingId: viewModel.id))
-                .onDisappear {
-                    viewModel.fetchShopping()
-                }
+            .onDisappear {
+                viewModel.fetchShopping()
+            }
+        }
+        .alert(isPresented: $viewModel.alertFlag) {
+            switch viewModel.alertType {
+            case .success:
+                Alert(title: Text(viewModel.alertMessage),
+                      dismissButton: .default(Text(R.string.button.close()),
+                                              action: { dismiss() }
+                                             )
+                )
+            case .error:
+                Alert(title: Text(viewModel.alertMessage),
+                      dismissButton: .default(Text(R.string.button.close()))
+                )
+            case .confirm:
+                Alert(title: Text(viewModel.alertMessage), primaryButton: .destructive(Text(R.string.button.delete()), action: {
+                    Task {
+                        viewModel.delete()
+                        viewModel.setAlert(type: .success, message: R.string.alertMessage.deleted())
+                    }
+                }),
+                      secondaryButton: .cancel(Text(R.string.button.cancel()))
+                )
+            }
         }
     }
 }
